requirements.txt
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.0
django-allauth==0.43.0
django-crispy-forms==1.10.0
django-environ==0.4.5
django-filter==2.4.0
django-import-export==2.4.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
idna==2.10
importlib-metadata==3.0.0
jdcal==1.4.1
Markdown==3.3.3
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.0.1
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
urllib3==1.26.2
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.0
